pyproject.toml
[tool.poetry]
name = "mirahezewikiauthbot"
version = "0.1.0"
description = ""
authors = ["pixDeVl <93949551+pixDeVl@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
flask = "^3.0.2"
mwoauth = "^0.4.0"
sqlalchemy = "^2.0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
